continuing development on unfinished body parts. change to base to address error where transform string would always be called on themselves. bodypartbase function modified to fix this. now works slightly differently, and is named differently from behavior. all existing classes updated to match this behavior.

+++ BodyPartsFinished/Nipples.cs
@@ -79,7 +79,7 @@ namespace CoC.BodyParts
 			return new Nipples(flags);
 		}
 
-		public static Nipples GenerateWithLength(PiercingFlags flags, int nippleLength)
+		public static Nipples GenerateWithLength(PiercingFlags flags, float nippleLength)
 		{
 			return new Nipples(flags)
 			{
@@ -87,7 +87,7 @@ namespace CoC.BodyParts
 			};
 		}
 
-		public static Nipples GenerateForceType(PiercingFlags flags, int nippleLength, NippleStatus status)
+		public static Nipples GenerateForceType(PiercingFlags flags, float nippleLength, NippleStatus status)
 		{
 			return new Nipples(flags)
 			{
